feat: Format text and change color scheme
Modify text colors to make it more readible with the background,
and refactor the function that populates the article body text
to replace "\r\n\r\n" with "<br /><br />" and replace the
remaining "\r\n" with " ". This ensures continuous flow of text
without the unwanted line breaks.

Modify the color scheme to use purple primary and primary dark
colors and light green accent color.