feat: Add theme coloring to Article Detail Metabar
Modify the metabar to have background inline with the app theme.

Apply the "Up" button animation to the entire container instead
of just the button.

Change the onClick(..) for the Up button to call onBackPressed()
instead of onSupportNavigateUp() to prevent data refesh every
time user navigates back to the homepage.